chore(tools): split pty tool module
