+++ src/tools/pty/output.rs
@@ -1,0 +1,19 @@
+use std::io::SeekFrom;
+use std::path::Path;
+
+use rara_tools::tool::ToolError;
+use tokio::io::{AsyncReadExt, AsyncSeekExt};
+
+pub(crate) async fn read_output_tail(path: &Path, max_bytes: usize) -> Result<String, ToolError> {
+    let mut file = match tokio::fs::File::open(path).await {
+        Ok(file) => file,
+        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(String::new()),
+        Err(err) => return Err(err.into()),
+    };
+    let file_len = file.metadata().await?.len();
+    let start = file_len.saturating_sub(max_bytes as u64);
+    file.seek(SeekFrom::Start(start)).await?;
+    let mut bytes = Vec::with_capacity(max_bytes.min(file_len as usize));
+    file.read_to_end(&mut bytes).await?;
+    Ok(String::from_utf8_lossy(&bytes).into_owned())
+}
